Renamed folders, fixed noduplicatesource so it works with any arbitrary attribute but defaults to src and has a function for src/id
