+++ CaptureRenderTagHelper.Benchmarks/Benchmarks/ScriptCaptureTagBenchmark.cs
