+++ CaptureRenderTagHelper.Benchmarks/Benchmarks/ScriptRenderTagBenchmark.cs
