+++ CaptureRenderTagHelper.Sample/CaptureRenderTagHelper.Sample.csproj
