+++ CaptureRenderTagHelper.Sample/Models/DisplayModel.cs
