+++ CaptureRenderTagHelper.Sample/Models/ViewModel.cs
