+++ CaptureRenderTagHelper.Sample/Views/Shared/DisplayTemplates/DisplayModel.cshtml
