+++ CaptureRenderTagHelper.Sample/Views/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using CaptureRenderTagHelper.Sample
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@addTagHelper *, CaptureRenderTagHelper
