+++ CaptureRenderTagHelper.Sample/appsettings.json
@@ -1,0 +1,7 @@
+﻿{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Warning"
+    }
+  }
+}
