+++ CaptureRenderTagHelper.sln
@@ -3,18 +3,18 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27512.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CaptureRenderTagHelper", "ScriptCaptureTagHelper\CaptureRenderTagHelper.csproj", "{C4498188-CB12-436B-B16F-4FEEE77440D6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CaptureRenderTagHelper", "CaptureRenderTagHelper\CaptureRenderTagHelper.csproj", "{C4498188-CB12-436B-B16F-4FEEE77440D6}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{53D5411B-E1D2-4AE0-BC1E-3BB00042E657}"
 	ProjectSection(SolutionItems) = preProject
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CaptureRenderTagHelper.Sample", "ScriptCaptureTagHelper.Sample\CaptureRenderTagHelper.Sample.csproj", "{C0A699A0-1BBF-4A63-8DEE-3959D1145B5A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CaptureRenderTagHelper.Sample", "CaptureRenderTagHelper.Sample\CaptureRenderTagHelper.Sample.csproj", "{C0A699A0-1BBF-4A63-8DEE-3959D1145B5A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CaptureRenderTagHelper.UnitTests", "ScriptCaptureTagHelper.UnitTests\CaptureRenderTagHelper.UnitTests.csproj", "{B25A8143-90B1-4354-B61F-DF5CFE17A218}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CaptureRenderTagHelper.UnitTests", "CaptureRenderTagHelper.UnitTests\CaptureRenderTagHelper.UnitTests.csproj", "{B25A8143-90B1-4354-B61F-DF5CFE17A218}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CaptureRenderTagHelper.Benchmarks", "ScriptCaptureTagHelper.Benchmarks\CaptureRenderTagHelper.Benchmarks.csproj", "{095C309B-28E3-4445-B31F-393730FFB93C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CaptureRenderTagHelper.Benchmarks", "CaptureRenderTagHelper.Benchmarks\CaptureRenderTagHelper.Benchmarks.csproj", "{095C309B-28E3-4445-B31F-393730FFB93C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
