--- ScriptCaptureTagHelper.Sample/Views/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using CaptureRenderTagHelper.Sample
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@addTagHelper *, CaptureRenderTagHelper
