--- ScriptCaptureTagHelper.Sample/appsettings.json
@@ -1,7 +1,0 @@
-﻿{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Warning"
-    }
-  }
-}
